requirements.txt
# aiofiles==22.1.0
# aiohttp==3.8.4
# aiosignal==1.3.1
# aiosqlite==0.18.0
# anyio==3.6.2
# appnope==0.1.3
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.2.3
# asttokens==2.2.1
# async-timeout==4.0.2
# attrs==22.2.0
# Babel==2.12.1
# backcall==0.2.0
# base58==2.1.1
# beautifulsoup4==4.11.2
# bitarray==2.7.3
# bleach==6.0.0
# certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==3.1.0
# comm==0.1.2
# cytoolz==0.12.1
# debugpy==1.6.6
# decorator==5.1.1
# defusedxml==0.7.1
# eth-abi==2.2.0
# eth-account==0.5.9
# eth-hash==0.5.1
# eth-keyfile==0.5.1
# eth-keys==0.3.4
# eth-rlp==0.2.1
# eth-typing==2.3.0
# eth-utils==1.9.5
# executing==1.2.0
# fastjsonschema==2.16.3
# fqdn==1.5.1
# frozenlist==1.3.3
# hexbytes==0.3.0
# idna==3.4
# importlib-metadata==6.0.0
# ipfshttpclient==0.8.0a2
# ipykernel==6.21.3
# ipython==8.11.0
# ipython-genutils==0.2.0
# isoduration==20.11.0
# jedi==0.18.2
# Jinja2==3.1.2
# json5==0.9.11
# jsonpointer==2.3
# jsonschema==4.17.3
# jupyter-events==0.6.3
# jupyter-ydoc==0.2.3
# jupyter_client==8.0.3
# jupyter_core==5.2.0
# jupyter_server==2.4.0
# jupyter_server_fileid==0.8.0
# jupyter_server_terminals==0.4.4
# jupyter_server_ydoc==0.6.1
# jupyterlab==3.6.1
# jupyterlab-pygments==0.2.2
# jupyterlab_server==2.20.0
# lru-dict==1.1.8
# MarkupSafe==2.1.2
# matplotlib-inline==0.1.6
# mistune==2.0.5
# multiaddr==0.0.9
# multidict==6.0.4
# nbclassic==0.5.3
# nbclient==0.7.2
# nbconvert==7.2.9
# nbformat==5.7.3
# nest-asyncio==1.5.6
# netaddr==0.8.0
# notebook==6.5.3
# notebook_shim==0.2.2
# numpy==1.24.2
# opencv-python==4.7.0.72
# packaging==23.0
# pandocfilters==1.5.0
# parsimonious==0.8.1
# parso==0.8.3
# pexpect==4.8.0
# pickleshare==0.7.5
# Pillow==9.4.0
# platformdirs==3.1.1
# prometheus-client==0.16.0
# prompt-toolkit==3.0.38
# protobuf==3.19.5
# psutil==5.9.4
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pycparser==2.21
# pycryptodome==3.17
# Pygments==2.14.0
# pyrsistent==0.19.3
# python-dateutil==2.8.2
# python-json-logger==2.0.7
# PyYAML==6.0
# pyzmq==25.0.0
# requests==2.28.2
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rlp==2.0.1
# Send2Trash==1.8.0
# six==1.16.0
# sniffio==1.3.0
# soupsieve==2.4
# stack-data==0.6.2
# terminado==0.17.1
# tinycss2==1.2.1
# tomli==2.0.1
# toolz==0.12.0
# tornado==6.2
# traitlets==5.9.0
# uri-template==1.2.0
# urllib3==1.26.15
# varint==1.0.2
# wcwidth==0.2.6
# web3==5.31.3
# webcolors==1.12
# webencodings==0.5.1
# websocket-client==1.5.1
# websockets==9.1
# y-py==0.5.9
# yarl==1.8.2
# ypy-websocket==0.8.2
# zipp==3.15.0
# pymerkle==4.0.0
#git+https://github.com/GraphitiLabs/zkaptcha-generator@69fa57d55bfff08f05ff50ebb53dd81360b868b6#egg=zkaptcha
pysha3
web3
eth-brownie
numpy
pandas
opencv-python
jupyterlab
pillow
git+https://github.com/magidandrew/pymerkle@master